Versioning and rudimentary timing
A rudimentary timing scheme was introduced though for testing various methods a more fine-grained approach is needed.

A new file is created when running "source" providing insight into the source created.
Data stored to this file includes number of files found.
The most important addition is a version number so that when changes are made to the source functionality inconsistencies are found by design.